move all vendor utils into vendorUtils
